% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.geefit.R
\name{pseudo.geefit}
\alias{pseudo.geefit}
\title{Function that computes makes GEE model fit}
\usage{
pseudo.geefit(pseudodata, covar_names)
}
\arguments{
\item{pseudodata}{Data set containing pseudo-observations. Expecting output from pseudo.twodim()}

\item{covar_names}{Vector with covariate names to be found in "pseudodata". E.g. covar_names = c("Z", "Z1")}
}
\description{
This function fits a GEE model based bivariate pseudo-observations of the marginal mean function and
the survival probability as returned by pseudo.twodim()
}
\examples{
pseudo.geefit()
}
\keyword{bivarpseudo}
