% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.threedim.R
\name{pseudo.threedim}
\alias{pseudo.threedim}
\title{Function that computes 3-dim pseudo-observations}
\usage{
pseudo.threedim(tstart, tstop, status, covar_names, id, tk, data, deathtype)
}
\arguments{
\item{tstart}{Start time - expecting counting process notation}

\item{tstop}{Stop time - expecting counting process notation}

\item{status}{Status variable (0 = censoring, 1 = recurrent event, 2 = death)}

\item{covar_names}{Vector containing names of covariates intended for further analysis}

\item{id}{ID variable for subject}

\item{tk}{Vector of time points to calculate pseudo-observations at}

\item{data}{Data set which contains variables of interest}

\item{deathtype}{Type of death (cause 1 or cause 2)}

\item{Z}{Covariate(s), Z}
}
\description{
This function computes 3-dimensional pseudo-observations of the marginal mean function and cumulative incidences of death causes 1 and 2
}
\examples{
pseudo.threedim()
}
\keyword{recurrentpseudo}
