% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.onedim.R
\name{pseudo.onedim}
\alias{pseudo.onedim}
\title{Function that computes univariate pseudo-observations}
\usage{
pseudo.onedim(tstart, tstop, status, covar_names, id, tk, data)
}
\arguments{
\item{tstart}{Start time - expecting counting process notation}

\item{tstop}{Stop time - expecting counting process notation}

\item{status}{Status variable (0 = censoring, 1 = recurrent event, 2 = death)}

\item{covar_names}{Vector containing names of covariates intended for further analysis}

\item{id}{ID variable for subject}

\item{tk}{Vector of time points to calculate pseudo-observations at}

\item{data}{Data set which contains variables of interest}

\item{Z}{Covariate(s), Z}
}
\description{
This function computes univariate pseudo-observations of the marginal mean function (in the presence of terminal events)
}
\examples{
pseudo.onedim()
}
\keyword{recurrentpseudo}
