% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.surv_cif_mu_est.R
\name{pseudo.surv_cif_mu_est}
\alias{pseudo.surv_cif_mu_est}
\title{Help function parsed inside pseudo.threedim()}
\usage{
pseudo.surv_cif_mu_est(inputdata, tstart, tstop, status, deathtype, id)
}
\arguments{
\item{inputdata}{Data set which contains variables of interest}

\item{tstart}{Start time - expecting counting process notation}

\item{tstop}{Stop time - expecting counting process notation}

\item{status}{Status variable (0 = censoring, 1 = recurrent event, 2 = death)}

\item{deathtype}{Type of death (cause 1 or cause 2)}

\item{id}{ID variable for subject}
}
\description{
Help function parsed inside pseudo.threedim()
}
\examples{
# Example: Bladder cancer data from survival package
require(survival)

# Add one extra day for the two patients with start=stop=0
# subset(bladder1, stop <= start)
bladder1[bladder1$id == 1, "stop"] <- 1
bladder1[bladder1$id == 49, "stop"] <- 1

# Restrict the data to placebo and thiotepa
bladdersub <- subset(bladder1, treatment \%in\% c("placebo", "thiotepa"))

# Make treatment variable two-level factor
bladdersub$Z <- as.factor(ifelse(bladdersub$treatment == "placebo", 0, 1))
levels(bladdersub$Z) <- c("placebo", "thiotepa")
head(bladdersub)

# Add deathtype variable to bladder data
# Deathtype = 1 (bladder disease death), deathtype = 2 (other death reason)
bladdersub$deathtype <- with(bladdersub, ifelse(status == 2, 1, ifelse(status == 3, 2, 0)))
table(bladdersub$deathtype, bladdersub$status)

#' # Make a three level status variable
bladdersub$status <- ifelse(bladdersub$status \%in\% c(2, 3), 2, bladdersub$status)


bladdersub$tstart <- bladdersub$start
bladdersub$tstop <- bladdersub$stop

# Estimate mu and cif1 and cif2 on data
pseudo.surv_cif_mu_est(inputdata = bladdersub,
                       tstart = bladdersub$start,
                       tstop = bladdersub$stop,
                       status = bladdersub$status,
                       deathtype = bladdersub$deathtype,
                       id = bladdersub$id)
}
\keyword{recurrentpseudo}
