% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.surv_mu_est.R
\name{pseudo.surv_mu_est}
\alias{pseudo.surv_mu_est}
\title{Help function parsed inside pseudo.twodim()}
\usage{
pseudo.surv_mu_est(inputdata, tstart, tstop, status, id)
}
\arguments{
\item{inputdata}{Data set which contains variables of interest}

\item{tstart}{Start time - expecting counting process notation}

\item{tstop}{Stop time - expecting counting process notation}

\item{status}{Status variable (0 = censoring, 1 = recurrent event, 2 = death)}

\item{id}{ID variable for subject}
}
\description{
Help function parsed inside pseudo.twodim()
}
\examples{
pseudo.surv_mu_est()
}
\keyword{bivarpseudo}
